backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
firebase-admin>=6.6.0
google-genai>=1.12.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.28.0
functions-framework>=3.8.0
